// src/scripts/interfaces/vehicle-service-info-data.interface.ts
import {VehicleServicePriceTable} from "@/scripts/interfaces/vehicle-service-price-table.interface";

export interface VehicleServiceInfoDataInterface {
    productCount: number;
    currentProductPrice: number;
    vehiclePrice: number;
    vehicleModelName: string;
    primaryColor: number;
    secondaryColor: number;
    vehicleDamagePercentage: number;
    priceTable: VehicleServicePriceTable;
}